Fixed root node not being a real node/Fixed node absolute lcn
The root node was pre-created and therefore not handled by the
directory handler. A dummy node is now used for this task and
so the real root node can be created by the directory handler.

The absolute lcn of a directory node had a trailing slash and
therefore the last part of the absolute lcn did not match the
lcn. This is now corrected.