Huge update to the node create-update-delete cycle
The previous tree update and write cycle depended on the knowledge
that only one such cycle is done during one render run. This made
it hard to add nodes to tree during the write phase of the cycle
and allow for the correct behaviour. The cycle implementation has
been changed sothat the SourceHandler#render call can now be done
many times during one webgen run.

This changes necessitated some changes in the Node and Path classes
that are incompatible with previous versions.