Special characters in paths are now allowed
The usage of umlauts and spaces is now allowed in source and output paths.
However, it may depend on the web server configuration if this works in
each case. Also note that it umlauts are just percent encoded, so if a
umlaut is in ISO-8859-1, it will get one escape sequence, but if it is
in UTF-8, it probably gets two escape sequences.