Implemented the item tracker extension manager
The ItemTracker extension manager is used to add items as dependencies
to a node and to check if a node or one of the items it depends on has
changed. Each managed extension is responsible for tracking a certain
item.

Currently, there are two item tracker extensions: one for tracking the
content of a node and one for tracking the meta information (either the
whole meta info hash or only a specific key) of a node.