Refactored Node#resolve! method
Since the Node class should not depend on website internals, the
Node#resolve! method was removed and an additional parameter
msg_on_failure to Node#resolve added. If this parameter is set
to true, the message :node_resolution_failed is dispatched.

The former functionality of Node#resolve! is now in a listener
defined by the ItemTracker::MissingNode extension where it belongs.