Add possibility to specify different path when routing to a node
Sometimes it is wanted that a node gets written out to a specific
destination path but when routing to this node another path should
be used. This can now be accomplished with the routing_path meta
information key.

One situation where this is useful is when one wants to create
"nice URLs" where normal pages are represented as directory index
files.