Require date library to avoid problems loading the cache file
Sometimes Date objects are in the cache, e.g. when the created_at meta
information is set to 2010-01-10. This *seems* to depend on the actually
used Ruby version and on the actually loaded libraries.

Since webgen doesn't require the date library anywhere, there is an
error when loading the cache due to the missing class Date.

Fix this by loading the date library.